Revert "Move custom `OmegaConf` resolvers to be discoverable plugin, instead of hooked into the `Runner` (#163)"
This reverts commit 2f7d840f1ac29f7ad444636a54f13dc05712ed6f.
