+++ hydra_plugins/searchpath/vital.py
@@ -17,4 +17,3 @@ class VitalSearchPathPlugin(SearchPathPlugin):  # noqa: D101
         # Remember to verify the config is packaged properly (build sdist and look inside,
         # and verify MANIFEST.in is correct).
         search_path.append(provider="vital-searchpath-plugin", path="pkg://vital.config")
-        search_path.append(provider="vital-plugins", path="pkg://hydra_plugins")
